BioPath/backend/requirements.txt
asgiref==3.5.2
tzdata==2022.2
Django==4.1.1
djangorestframework==3.13.1
drf_yasg==1.21.3
sqlparse==0.4.2
psycopg2-binary>=2.8
Pillow
django-cors-headers
